chore: remove old seed data
