+++ packages/hono-backend/package.json
@@ -12,7 +12,7 @@
         "db:push": "drizzle-kit push",
         "db:studio": "drizzle-kit studio",
         "db:drop": "drizzle-kit drop",
-        "db:seed": "bun run src/db/seed/seed-new.ts",
+        "db:seed": "bun run src/db/seed/index.ts",
         "db:seed:refresh": "bun run src/db/seed/seed-refresh.ts",
         "db:dangerously-reset-do-not-run-in-production": "psql $DATABASE_URL -c 'DROP SCHEMA public CASCADE; DROP SCHEMA drizzle CASCADE; CREATE SCHEMA public;' && drizzle-kit migrate && bun run src/db/seed/index.ts"
     },
